Renaming the plugin to resource sharing.
