+++ resource_sharing/gui/manage_dialog.py
