+++ resource_sharing/gui/ui/download_dialog.ui
