+++ resource_sharing/gui/ui/manage_repository.ui
