+++ resource_sharing/gui/ui/resource_sharing_dialog_base.ui
