+++ resource_sharing/network_manager.py
