+++ resource_sharing/repository_handler/bitbucket_handler.py
@@ -1,29 +1,29 @@
 # coding=utf-8
-from symbology_sharing.repository_handler.remote_git_handler import (
+from resource_sharing.repository_handler.remote_git_handler import (
     RemoteGitHandler)
 
 
-class GithubHandler(RemoteGitHandler):
-    """Handler class for Github Repository."""
+class BitBucketHandler(RemoteGitHandler):
+    """Handler class for Bitbucket Repository."""
     IS_DISABLED = False
 
     def __init__(self, url=None):
         RemoteGitHandler.__init__(self, url)
 
     def can_handle(self):
-        if self.git_platform == 'github':
+        if self.git_platform == 'bitbucket':
             return True
         return False
 
     @property
     def metadata_url(self):
-        return 'https://raw.githubusercontent.com/%s/%s/master/%s' % (
+        return 'https://bitbucket.org/%s/%s/raw/master/%s' % (
             self.git_owner,
             self.git_repository,
             self.METADATA_FILE)
 
     def preview_url(self, collection_name, filename):
-        return 'https://raw.githubusercontent.com/%s/%s/master/collections/%s/preview/%s' % (
+        return 'https://bitbucket.org/%s/%s/raw/master/collections/%s/preview/%s' % (
             self.git_owner,
             self.git_repository,
             collection_name,
