+++ resource_sharing/resource_handler/processing_handler.py
@@ -7,8 +7,8 @@ from processing.script.ScriptAlgorithm import ScriptAlgorithm
 from processing.script.WrongScriptException import WrongScriptException
 from processing.script.ScriptUtils import ScriptUtils
 
-from symbology_sharing.resource_handler.base import BaseResourceHandler
-from symbology_sharing.utilities import qgis_version
+from resource_sharing.resource_handler.base import BaseResourceHandler
+from resource_sharing.utilities import qgis_version
 
 
 class ProcessingScriptHandler(BaseResourceHandler):
