+++ resource_sharing/resource_handler/style_handler.py
@@ -2,8 +2,8 @@
 import os
 import fnmatch
 
-from symbology_sharing.resource_handler.base import BaseResourceHandler
-from symbology_sharing.resource_handler.symbol_resolver_mixin import SymbolResolverMixin
+from resource_sharing.resource_handler.base import BaseResourceHandler
+from resource_sharing.resource_handler.symbol_resolver_mixin import SymbolResolverMixin
 
 
 class StyleResourceHandler(BaseResourceHandler, SymbolResolverMixin):
