+++ resource_sharing/symbol_xml_extractor.py
